+++ ouzel/graphics/opengl/emscripten/OGLRenderDeviceEm.cpp
@@ -23,7 +23,7 @@ namespace ouzel
         }
 
         void OGLRenderDeviceEm::init(Window* newWindow,
-                                     const Size2<float>& newSize,
+                                     const Size2<uint32_t>& newSize,
                                      uint32_t newSampleCount,
                                      Texture::Filter newTextureFilter,
                                      uint32_t newMaxAnisotropy,
